poisson probability

+++ fdbrpc/include/fdbrpc/simulator.h
@@ -83,7 +83,7 @@ public:
 		bool rebooting;
 		std::vector<flowGlobalType> globals;
 
-		int buggifyDelayCount = 0;
+		uint8_t buggifyDelayCount = 0;
 
 		INetworkConnections* network;
